third_party/blink/web_tests/html5lib/generated/run-menuitem-element-data.html

<!DOCTYPE html>
<script>
var test_files = [ '../resources/menuitem-element.dat' ]
</script>
<script src="../../resources/dump-as-markup.js"></script>
<script>window.forceDataURLs = true;</script>
<script src="../resources/runner.js"></script>
