third_party/blink/web_tests/html5lib/generated/run-menuitem-element-data-expected.txt

../resources/menuitem-element.dat:
14

Test 14 of 20 in ../resources/menuitem-element.dat failed. Input:
<!DOCTYPE html><select><menuitem></select>
Got:
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <select>
|       <menuitem>
Expected:
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <select>
